# feat: dramatically simplify AI instructions - 80% reduction for better consistency